Implemented Storage, Server-side Caching, Export

+++ src/services/postgres/PlaylistsService.js
@@ -7,9 +7,10 @@ const NotFoundError = require('../../exceptions/NotFoundError')
 const AuthorizationError = require('../../exceptions/AuthorizationError')
 
 class PlaylistsService {
-  constructor (collaborationService) {
+  constructor (collaborationService, cacheService) {
     this._pool = new Pool()
     this._collaborationService = collaborationService
+    this._cacheService = cacheService
   }
 
   async addPlaylist ({ name, owner }) {
@@ -26,25 +27,48 @@ class PlaylistsService {
       throw new InvariantError('Playlist gagal ditambahkan')
     }
 
+    await this._cacheService.delete(`playlists:${owner}`)
     return result.rows[0].id
   }
 
-  async getPlaylists (owner) {
-    const query = {
-      text: `SELECT playlists.id, playlists.name, users.username FROM playlists
-      JOIN users ON playlists.owner=users.id
-      LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
-      WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
-      values: [owner]
-    }
+  // const query = {
+  //   text: `SELECT playlists.id, playlists.name, users.username FROM playlists
+  //   JOIN users ON playlists.owner=users.id
+  //   LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
+  //   WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
+  //   values: [owner]
+  // }
 
-    const result = await this._pool.query(query)
-    return result.rows.map(mapPlaylistModel)
+  // const result = await this._pool.query(query)
+  // return result.rows.map(mapPlaylistModel)
+  async getPlaylists (owner) {
+    try {
+      // mendapatkan playlist dari cache
+      const result = await this._cacheService.get(`playlists:${owner}`)
+      return JSON.parse(result)
+    } catch (error) {
+      // bila gagal, diteruskan dengan mendapatkan playlists dari database
+      const query = {
+        text: `SELECT playlists.id, playlists.name, users.username FROM playlists
+        JOIN users ON playlists.owner=users.id
+        LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
+        WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
+        values: [owner]
+      }
+
+      const result = await this._pool.query(query)
+      const mappedResult = result.rows.map(mapPlaylistModel)
+
+      // playlist akan disimpan pada cache sebelum fungsi getPlaylist dikembalikan
+      await this._cacheService.set(`playlists:${owner}`, JSON.stringify(mappedResult))
+
+      return mappedResult
+    }
   }
 
   async deletePlaylistById (id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id]
     }
 
@@ -53,6 +77,9 @@ class PlaylistsService {
     if (!result.rows.length) {
       throw new NotFoundError('Playlist gagal dihapus. Id tidak ditemukan')
     }
+
+    const { owner } = result.rows[0]
+    await this._cacheService.delete(`playlists:${owner}`)
   }
 
   async verifyPlaylistOwner (id, owner) {
